Created a ChooseFacility.cs that replaced all the other choose facility methods
