+++ src/Actions/ChooseFacility.cs
@@ -12,6 +12,9 @@ namespace Trestlebridge.Actions
     {
         public static void CollectInput(Farm farm, List<IFacility<IResource>> facilities, IResource resource)
         {
+
+            Console.Clear();
+
             List<IFacility<IResource>> ListOfFacilites = facilities.Where(facility => facility.GetCount < facility.Capacity).ToList();
 
             List<IFacility<IResource>> copy = ListOfFacilites;
@@ -25,13 +28,21 @@ namespace Trestlebridge.Actions
             else
             {
 
-                ListOfFacilites.ForEach(facility => Console.WriteLine($"{copy.IndexOf(facility) + 1}) {facility}"));
+                ListOfFacilites.ForEach(facility => {
+                    Console.WriteLine($"{copy.IndexOf(facility) + 1}) {facility}");
+                    Console.WriteLine($"List of resources in facility {copy.IndexOf(facility) + 1}:");
+                    foreach(KeyValuePair<string, int> kvp in facility.GetTypeCount())
+                    {
+                        Console.Write($"{kvp.Key}: {kvp.Value} ");
+                    }
+                    Console.WriteLine();
+                } );
 
 
                 Console.WriteLine ();
 
                 // How can I output the type of animal chosen here?
-                Console.WriteLine ($"Place the animal where?");
+                Console.WriteLine ($"Place the resource where?");
 
                 Console.Write ("> ");
 
@@ -47,6 +58,7 @@ namespace Trestlebridge.Actions
                         {
                             facilityO = ListOfFacilites[choice - 1];
                             ListOfFacilites[choice - 1].AddResource(resource);
+                            Console.Clear();
                             Console.WriteLine(ListOfFacilites[choice - 1]);
                             Console.WriteLine();
                             Console.WriteLine("Press enter to return to the main menu.");
@@ -67,6 +79,100 @@ namespace Trestlebridge.Actions
             }
 
             
+        }
+        public static void CollectInput(Farm farm, List<IFacility<IResource>> naturalFields, List<IFacility<IResource>> plowedFields, IResource resource)
+        {
+            Console.Clear();
+
+
+
+            //List of available Natural Fields
+            List<IFacility<IResource>> AvailableNattyFields = new List<IFacility<IResource>>();
+
+            for (int i = 0; i < farm.NaturalFields.Count; i++)
+            {
+                if (farm.NaturalFields[i].GetCount < farm.NaturalFields[i].Capacity)
+                {
+                    AvailableNattyFields.Add(farm.NaturalFields[i]);
+                }
+            }
+            //List of available Plowable fields
+            List<IFacility<IResource>> AvailablePlowwyFields = new List<IFacility<IResource>>();
+
+            for (int i = 0; i < farm.PlowedFields.Count; i++)
+            {
+                if (farm.PlowedFields[i].GetCount < farm.PlowedFields[i].Capacity)
+                {
+                    AvailablePlowwyFields.Add(farm.PlowedFields[i]);
+                }
+            }
+            //List of available fields of both types: Shared with both IFacility and IResource
+            List<IFacility<IResource>> fields = AvailableNattyFields.Concat(AvailablePlowwyFields).ToList();
+
+
+            if (fields.Count == 0)
+            {
+                Console.WriteLine("There are no fields with enough space available! Please create either a natural field or a plowed field. Press any key to continue.");
+                Console.Read();
+            }
+            else
+            {
+                for (int i = 0; i < fields.Count; i++)
+                {
+
+
+                    Console.WriteLine($"{i + 1}. {fields[i].GetType().Name} ({fields[i].GetCount} seeds),");
+
+                    foreach (KeyValuePair<string, int> seeds in fields[i].GetTypeCount())
+                    {
+
+                        Console.Write($" ({seeds.Key}: {seeds.Value}) ");
+
+                    }
+
+                    Console.WriteLine();
+                }
+                Console.WriteLine();
+                Console.WriteLine($"Place the seeds where?");
+
+                Console.Write("> ");
+
+
+                int choice = 0;
+                object field = null;
+
+                while (choice == 0 || field == null)
+                {
+                    try
+                    {
+                        choice = Int32.Parse(Console.ReadLine());
+
+                        try
+                        {
+                            field = fields[choice - 1];
+                            fields[choice - 1].AddResource(resource);
+                            Console.WriteLine(fields[choice - 1]);
+                            Console.WriteLine();
+                            Console.WriteLine("Press any key to return to the main menu.");
+                            Console.Read();
+                        }
+                        catch (ArgumentOutOfRangeException)
+                        {
+                            Console.WriteLine("Invalid Selection!");
+                            Console.WriteLine($"Place the seeds where?");
+                        }
+
+                    }
+                    catch (FormatException)
+                    {
+
+                        Console.WriteLine("Please enter a number!");
+                        Console.WriteLine($"Place the seeds where?");
+                    }
+
+
+                }
+            }
         }
     }
 }
